+++ projects/ez-vid-ang/src/lib/core/directives/keyboard-shortcuts.spec.ts
@@ -1,0 +1,8 @@
+import { EvaKeyboardShortcuts } from './keyboard-shortcuts';
+
+describe('KeyboardShortcuts', () => {
+  it('should create an instance', () => {
+    const directive = new EvaKeyboardShortcuts();
+    expect(directive).toBeTruthy();
+  });
+});